requirements.txt
Flask
Flask-Cors
python-dotenv
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-web
